Evolving Cow-things
The cows now have a random chance of receiving a slight change in
several traits that decide their survivability. There are now
meat-eating cows that have a 1/500 chance of evolving from a normal cow,
they serve to reduce the population as well as provide a diverse
simulation. The color of the cows now reflects their traits.